backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.3.0
anthropic>=0.39.0
google-generativeai>=0.8.0
psycopg2-binary>=2.9.9



